changing to float

+++ src/br/edu/uniaeso/q7/Rhombus.java
@@ -1,21 +1,21 @@
 package br.edu.uniaeso.q7;
 
 public class Rhombus {
-    private int diagonal1, diagonal2;
+    private float diagonal1, diagonal2;
 
-    public int getDiagonal1() {
+    public float getDiagonal1() {
         return diagonal1;
     }
 
-    public void setDiagonal1(int diagonal1) {
+    public void setDiagonal1(float diagonal1) {
         this.diagonal1 = diagonal1;
     }
 
-    public int getDiagonal2() {
+    public float getDiagonal2() {
         return diagonal2;
     }
 
-    public void setDiagonal2(int diagonal2) {
+    public void setDiagonal2(float diagonal2) {
         this.diagonal2 = diagonal2;
     }
 
@@ -28,4 +28,5 @@ public class Rhombus {
         setDiagonal1(0);
         setDiagonal2(0);
     }
+
 }
